requirements.txt
streamlit==1.37.0
python-dotenv
llama-index
llama-index-embeddings-huggingface
llama-index-llms-huggingface
